Simplify code and improve documentation on PasswordStrengthMeter

+++ project/app/src/main/java/com/example/jacob/tddb84project/PasswordStrengthMeter/algorithm/StrengthValidatorInterface.java
@@ -1,5 +1,9 @@
 package com.example.jacob.tddb84project.PasswordStrengthMeter.algorithm;
 
 public interface StrengthValidatorInterface {
-    float validate(String password);
+    /**
+     * Validation score
+     * Should return a value between 0-1
+     */
+    double validate(String password);
 }
